Add login functionality to BAseTest class
