+++ SeleniumTests/PiguLT/SearchForItem.cs
@@ -10,7 +10,6 @@ namespace SeleniumTests.PiguLT
         [Test]
         public void SearchItemInWebShop()
         {
-            SearchPage.ClickAllowAllCookies();
             SearchPage.EnterSearchPhrase(Strings.SearchItemString.ItemSearchPhrase1);
             SearchPage.ClickSearchButton();
             SearchPage.ClickFoundItemImageButton();
